Gaussian blurs: StackBlur wasn't actually operating in 16 bit
 * The original AGG code has a version that is supposedly 30 to 40
   percent slower, but works in any colorspace and with any blur
   radius. So this doesn't only fix jagged dark gradients in the
   blurs, but also fixes blurs with radius greater than 254 pixels.
 * The DropShadow filter now uses the transformation to transform the
   drop offset. One can thus also move the drop shadow via the
   regular Transform tool.

+++ src/model/snapshots/FilterDropShadowSnapshot.h
@@ -34,8 +34,8 @@ class FilterDropShadowSnapshot : public ObjectSnapshot {
 			float				fOffsetY;
 			float				fOpacity;
 			float				fLayoutedFilterRadius;
-			float				fLayoutedOffsetX;
-			float				fLayoutedOffsetY;
+			double				fLayoutedOffsetX;
+			double				fLayoutedOffsetY;
 			rgb_color			fColor;
 };
